requirements.txt
django==3.2.13
djangorestframework==3.12.4
django-cors-headers==3.11.0
psycopg2-binary==2.9.3
